Refactor reusable_resource and state_variable classes
This commit refactors the reusable_resource and state_variable classes by adding new functionality to the get_data() methods in the rr_flaw and sv_flaw classes. The get_data() methods now return JSON objects with additional information about the atoms involved in the flaws. This improves the API functionality and enhances the data representation of rr_flaw and sv_flaw objects.